Add conditional support for overriding secutiryContext in helm chart
It is conditional because when MIG is enabled we *have to* have the
following settings at a minimum:
```
  env:
    - name: NVIDIA_MIG_MONITOR_DEVICES
      value: all
  securityContext:
    capabilities:
      add:
        - SYS_ADMIN # or something else that adds this capability
```

With this patch, if a user overrides `securityContext: {}` then that
will take precedence. If they do not override `securityContext: {}` (and
a mig strategy other than `none` is set) it will be set as indicated
above.

Signed-off-by: Kevin Klues <kklues@nvidia.com>