Add files necessary to treat this as a helm repository
There is currently only a single repository called stable which will
house both pre-release and full release charts. Helm forces you to add a
--devel flag to your `helm install` commands in order to insrtall
pre-releases, so it's OK to include them in the asame repo as full
releases. The top level index is a symlink to the stable repository in
case we want to add others here in the future (e.g. nightly).

Signed-off-by: Kevin Klues <kklues@nvidia.com>