Use the launch / run timestamp as the gfd.timestamp label value
This change ensures that the nvidia.com/gfd.timestamp label is updated
only at start up and not with every loop iteration.

This means that the timestamp could be used as a useful signal for monitoring
the health of the application, as a changine timestamp would indicate that
the application is being restarted continuously.

The test that tests the looping / sleeping behaviour is also modified as the
timestamp is not available to be used as an indicator of changed file contents.

Signed-off-by: Evan Lezar <elezar@nvidia.com>