Switch to urfave/cli for handling of options
This change switches to using urfave/cli for handling of
command line arguments and envvars. This  allows for the
use of the config file as defined by the device plugin in a future
commit.

Note that this does change the order of preference for how options
are applied in that command line arguments take precedence over
environment variables.

Signed-off-by: Evan Lezar <elezar@nvidia.com>